+++ src/modules/show_meals.js
@@ -1,19 +1,8 @@
 /* eslint-disable no-useless-concat */
-const show = (arr) => {
-  const ProductList = document.querySelector('#meals-content');
-  const fetchLikes = async () => {
-    const response = await fetch(
-      'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/o8FdGibCYqeybnMSFY1p/likes/',
-    ).then((res) => res.json());
-    response.forEach((data) => {
-      const likeItem = ProductList.querySelector(`div[id="${data.item_id}"]`);
-      const li = likeItem.querySelector('#n-likes');
-      if (`${data.likes}` >= 2) { li.innerHTML = `${data.likes} ` + 'Likes'; } else {
-        li.innerHTML = `${data.likes} ` + 'Like';
-      }
-    });
-  };
+import fetchLikes from './newLike.js';
 
+const show = (arr) => {
+  fetchLikes();
   let html = '';
   // eslint-disable-next-line no-restricted-syntax
   for (const meal of arr) {
@@ -32,6 +21,7 @@ const show = (arr) => {
      </div>
         `;
   }
+
   const mainContent = document.getElementById('meals-content');
   mainContent.innerHTML = html;
   fetchLikes();
